Small cloud build service refactoring
