+++ lib/definitions/cloud-build-helper.d.ts
@@ -1,0 +1,7 @@
+declare interface ICloudBuildHelper {
+	getProvisionType(provisionData: IMobileProvisionData): string;
+	getCertificateInfo(certificatePath: string, certificatePassword: string): ICertificateInfo;
+	isReleaseConfiguration(buildConfiguration: string): boolean;
+	getMobileProvisionData(provisionPath: string): IMobileProvisionData;
+	zipProject(projectDir: string): Promise<string>;
+}
